fix(sim): fix euler angle int truncation caused by numpy dtype=int64
Root cause: set_euler([[0], [0], [0]]) created a numpy array with
dtype=int64 (integer literals). When step_fast() wrote float values
like atan2()=0.785 into euler[2][0], numpy silently truncated them
to integers (0.785 → 0, 2.356 → 2).

This caused euler[2][0] to appear as 0 for any yaw < 57.3° (1 rad),
which is why the camera yaw tracking failed — it was reading 0.

Fix: use dtype=float in np.array() for euler in both __init__ and
set_euler to ensure float storage regardless of input literals.

+++ simulator/vpython/core/physics.py
@@ -46,7 +46,7 @@ class rigidbody():
         # Pre-compute inverse inertia matrix (constant, avoids 8000 inv/s)
         # 慣性逆行列を事前計算（定数なので毎ステップの計算を回避）
         self.inertia_inv = np.linalg.inv(np.array(inersia))
-        self.euler = np.array(euler)
+        self.euler = np.array(euler, dtype=float)
         self.quat = self.euler2quat(euler)
         self.DCM = self.quat_dcm(self.quat)
         self.position = np.array(position)#np.zeros((3,1))
@@ -238,7 +238,7 @@ class rigidbody():
         self._update_dcm_cache(self.dcm)
 
     def set_euler(self, euler):
-        self.euler = np.array(euler)
+        self.euler = np.array(euler, dtype=float)
         self.quat = self.euler2quat(self.euler)
         self.dcm = self.euler_dcm(self.euler)
         self.s_q0 = float(self.quat[0][0])
